+++ operators/backface_select/__init__.py
@@ -7,6 +7,7 @@ allowing selection of visible geometry behind them without X-ray mode.
 
 import bpy
 from . import operator
+from . import paint_select
 
 
 _addon_keymaps = []
@@ -14,6 +15,7 @@ _addon_keymaps = []
 
 def register():
     operator.register()
+    paint_select.register()
 
     wm = bpy.context.window_manager
     kc = wm.keyconfigs.addon
@@ -66,6 +68,15 @@ def register():
     kmi.properties.loop = True
     _addon_keymaps.append((km, kmi))
 
+    # Ctrl+click (paint select)
+    kmi = km.keymap_items.new(
+        paint_select.LEVELDESIGN_OT_backface_paint_select.bl_idname,
+        'LEFTMOUSE', 'PRESS',
+        ctrl=True,
+        head=True
+    )
+    _addon_keymaps.append((km, kmi))
+
     # Object mode keymaps
     km_obj = kc.keymaps.new(name='Object Mode', space_type='EMPTY')
 
@@ -94,4 +105,5 @@ def unregister():
         km.keymap_items.remove(kmi)
     _addon_keymaps.clear()
 
+    paint_select.unregister()
     operator.unregister()