+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.2.5
 - Selecting, texture application, cube cut and box builder all now ignore culled backfaces
+- New select mode: PAINT SELECT. Hold ctrl, left mouse, and drag to add all vertices, edges, or faces crossed to the selection (by default overrides 'shortest path to')
+- Alt left mouse to apply textures now works hold left mouse and drag
 
 ## 1.2.4
 - Fix hotspot mapping orientation type icon sizing (probably)